# docs/openapi.yaml
openapi: 3.1.0
info:
  title: VelesDB API
  description: High-performance vector database for AI applications. Supports semantic search, HNSW indexing, and multiple distance metrics. Authentication is optional — when API keys are configured via VELESDB_API_KEYS, all endpoints except /health and /ready require a valid Bearer token.
  contact:
    name: VelesDB Team
    url: https://github.com/cyberlife-coder/VelesDB
  license:
    name: VelesDB Core License 1.0
    url: https://github.com/cyberlife-coder/VelesDB/blob/main/LICENSE
  version: 6.0.0
servers:
- url: /
  description: Local server
paths:
...
  /collections:
    get:
      tags:
      - collections
      summary: List all collections.
      operationId: list_collections
      responses:
        '200':
          description: List of collections
          content:
            application/json:
              schema:
                type: object
    post:
      tags:
      - collections
      summary: Create a new collection.
      operationId: create_collection
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCollectionRequest'
        required: true
      responses:
        '201':
          description: Collection created
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 'A body the server cannot deserialise: a field of the wrong type, or JSON that does not match the request schema. Axum''s extractor answers before the handler runs, so the payload is plain text, not an ErrorResponse.'
          content:
            text/plain:
              schema:
                type: string
  /collections/{name}:
    get:
      tags:
      - collections
      summary: Get collection information.
      operationId: get_collection
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Collection details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
      - collections
      summary: Delete a collection.
      operationId: delete_collection
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Collection deleted
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /collections/{name}/empty:
    get:
      tags:
      - collections
      summary: Check if a collection is empty.
      operationId: is_empty
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Empty status
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /collections/{name}/flush:
    post:
      tags:
      - collections
      summary: Flush pending changes to disk.
      operationId: flush_collection
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Flushed successfully
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Flush failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /collections/{name}/sanity:
    get:
      tags:
      - collections
      summary: Run a quick sanity check for onboarding and troubleshooting.
      operationId: collection_sanity
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Collection sanity status
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    CollectionResponse:
      type: object
      description: Response with collection information.
      required:
      - name
      - dimension
      - metric
      - point_count
      - storage_mode
      properties:
        dimension:
          type: integer
          description: Vector dimension.
          minimum: 0
        metric:
          type: string
          description: Distance metric.
        name:
          type: string
          description: Collection name.
        point_count:
          type: integer
          description: Number of points in the collection.
          minimum: 0
        storage_mode:
          type: string
          description: Storage mode (full, sq8, binary).
...
    CreateCollectionRequest:
      type: object
      description: Request to create a new collection.
      required:
      - name
      properties:
        async_index_builder:
          description: |-
            Async index builder configuration (Issue `#488` — Bulk Insert V2).

            When present, enables the `AsyncIndexBuilder` for deferred HNSW
            insertion during bulk import. Accepted as a free-form JSON object
            matching
            `velesdb_core::collection::streaming::AsyncIndexBuilderConfig`.
        collection_type:
          type: string
          description: 'Collection type: "vector" (default) or "`metadata_only`".'
          example: vector
        deferred_indexing:
          description: |-
            Deferred indexing configuration (`US-366`).

            When present, inserts are buffered in memory and batch-merged
            into the HNSW index when the buffer reaches `merge_threshold`.
            Accepted as a free-form JSON object matching
            `velesdb_core::collection::streaming::DeferredIndexerConfig`.
        dimension:
          type:
          - integer
          - 'null'
          description: Vector dimension (required for vector collections, ignored for `metadata_only`).
          example: 768
          minimum: 0
        graph_schema:
          description: |-
            Graph schema (only for `collection_type = "graph"`).

            Accepted as a free-form JSON object matching
            `velesdb_core::GraphSchema`. `GraphSchema::schemaless()` is
            used when the field is absent.
        hnsw_alpha:
          type:
          - number
          - 'null'
          format: float
          description: |-
            VAMANA alpha for neighbor diversification (default derived from dimension).

            Values > 1.0 bias `select_neighbors` toward diversity over proximity,
            producing a more navigable graph with better recall. The engine
            default is 1.2 (VAMANA paper recommendation). Set to 1.0 for strict
            nearest-neighbor selection.
          example: 1.2
        hnsw_ef_construction:
          type:
          - integer
          - 'null'
          description: HNSW `ef_construction` parameter (candidate list size during build).
          example: 400
          minimum: 0
        hnsw_m:
          type:
          - integer
          - 'null'
          description: HNSW M parameter (number of bi-directional links per node).
          example: 32
          minimum: 0
        hnsw_max_elements:
          type:
          - integer
          - 'null'
          description: |-
            Initial HNSW capacity (grows automatically if exceeded).

            Pre-sizing matters for bulk imports: the engine avoids repeated
            realloc-and-copy cycles when the final population is known upfront.
            Use this when migrating a large existing corpus into a fresh
            collection.
          example: 1000000
          minimum: 0
        metric:
          type: string
          description: Distance metric (cosine, euclidean, dot, hamming, jaccard).
          example: cosine
        name:
          type: string
          description: Collection name.
          example: documents
        pq_rescore_oversampling:
          type:
          - integer
          - 'null'
          format: int32
          description: |-
            PQ rescore oversampling factor (default 4).

            The search pipeline fetches `max(k * factor, k + 32)` candidates
            from HNSW and rescores them with full-precision ADC. Only
            meaningful for quantised storage modes (SQ8, PQ). `Some(0)` is
            treated as "disabled".
          example: 4
          minimum: 0
        storage_mode:
          type: string
          description: Storage mode (full, sq8, binary). Defaults to full.
          example: full
...
    ErrorResponse:
      type: object
      description: Error response.
      required:
      - error
      properties:
        code:
          type:
          - string
          - 'null'
          description: Structured error code (e.g., `"VELES-004"` for dimension mismatch).
        error:
          type: string
          description: Error message.
...
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
security:
- bearer_auth: []
...